Add author service
In addition also added author book linking for creation and update in book.
The idea is that Author ID will not be serialized to JSON but deserialized
only. While Author will be ignored all time. In addition to that author
service will be able to populate Author from ID in Book.

Signed-off-by: Imran M Yousuf <imyousuf@smartitengineering.com>